UPAR/requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.18.1
Pillow==7.1.2
